requirements.txt
gunicorn
requests
translate
djangorestframework
django
django-heroku